[UPDATE] split result handler
