+++ handler/http_handler.go
@@ -118,3 +118,89 @@ func (h *Handler) PurchaseHandler(w http.ResponseWriter, r *http.Request) {
 
 	http.Redirect(w, r, "/Purchase_page?id="+l.ID, http.StatusFound)
 }
+
+
+// handler for result
+func (h *Handler) ResultHandler(w http.ResponseWriter, r *http.Request) {
+    resp1, err := http.Get("https://lottery-dot-tenntenn-samples.appspot.com/result?id=" + r.FormValue("id"))
+    if err != nil {
+        const status = http.StatusInternalServerError
+        http.Error(w, http.StatusText(status), status)
+        return
+    }
+    defer resp1.Body.Close()
+
+    var result types.Result
+    if err := json.NewDecoder(resp1.Body).Decode(&result); err != nil {
+        const status = http.StatusInternalServerError
+        http.Error(w, http.StatusText(status), status)
+        return
+    }
+
+    resp2, err := http.Get("https://lottery-dot-tenntenn-samples.appspot.com/lottery?id=" + r.FormValue("id"))
+    if err != nil {
+        const status = http.StatusInternalServerError
+        http.Error(w, http.StatusText(status), status)
+        return
+    }
+    defer resp2.Body.Close()
+
+    var l types.Lottery
+    if err := json.NewDecoder(resp2.Body).Decode(&l); err != nil {
+        const status = http.StatusInternalServerError
+        http.Error(w, http.StatusText(status), status)
+        return
+    }
+
+    type winner struct {
+        Prize   *types.Prize
+        Numbers []string
+    }
+
+    data := struct {
+        types.Lottery
+        Winners map[string]*winner
+    }{
+        Lottery: l,
+        Winners: map[string]*winner{},
+    }
+
+    rows, err := h.DB.Query("SELECT number FROM Purchased WHERE lottery_id = ?", l.ID)
+    if err != nil {
+        const status = http.StatusInternalServerError
+        http.Error(w, http.StatusText(status), status)
+        return
+    }
+    for rows.Next() {
+        var number string
+        if err := rows.Scan(&number); err != nil {
+            const status = http.StatusInternalServerError
+            http.Error(w, http.StatusText(status), status)
+            return
+        }
+
+        for i := range result.Winners {
+            for _, n := range result.Winners[i].Numbers {
+                if number == n {
+                    prizeID := result.Winners[i].PrizeID
+                    if data.Winners[prizeID] == nil {
+                        for _, p := range l.Prizes {
+                            if p.ID == prizeID {
+                                data.Winners[prizeID] = &winner{
+                                    Prize: p,
+                                }
+                            }
+                        }
+                    }
+                    data.Winners[prizeID].Numbers = append(data.Winners[prizeID].Numbers, n)
+                }
+            }
+        }
+    }
+
+    if err := preview.ResultTmpl.Execute(w, data); err != nil {
+        const status = http.StatusInternalServerError
+        http.Error(w, http.StatusText(status), status)
+        return
+    }
+}
